+++ internal/resolver/router.go
@@ -11,6 +11,8 @@ import (
 // CachingRouter implements resolver for routers by uuid or name, caching the results
 type CachingRouter struct {
 	Cache[upcloud.Router]
+
+	Type string
 }
 
 var (
@@ -26,12 +28,15 @@ func (s *CachingRouter) Get(ctx context.Context, svc internal.AllServices) (Reso
 	}
 
 	for _, router := range routers.Routers {
+		if s.Type != "" && router.Type != s.Type {
+			continue
+		}
 		s.AddCached(router.UUID, router)
 	}
 
 	return func(arg string) Resolved {
 		rv := Resolved{Arg: arg}
-		for _, router := range routers.Routers {
+		for _, router := range s.cache {
 			rv.AddMatch(router.UUID, MatchTitle(arg, router.Name))
 			rv.AddMatch(router.UUID, MatchUUID(arg, router.UUID))
 		}